+++ lib/kalender.dart
@@ -49,7 +49,7 @@ export 'package:kalender/src/widgets/components/month_grid.dart';
 export 'package:kalender/src/widgets/components/multi_day_overlay.dart';
 export 'package:kalender/src/widgets/components/multi_day_overlay_portal.dart';
 export 'package:kalender/src/widgets/components/multi_day_overlay_portal_button.dart';
-export 'package:kalender/src/widgets/components/resize_handle_positioner.dart';
+export 'package:kalender/src/widgets/components/resize_handles.dart';
 export 'package:kalender/src/widgets/components/time_indicator.dart';
 export 'package:kalender/src/widgets/components/time_line.dart';
 export 'package:kalender/src/widgets/components/week_day_header.dart';